Change tests organization
Each "module", such as "writer", "reader", etc. will have a separate
test source file that is compiled as an executable. Then the `ctest`
command, or the `make test` command can be executed to run the tests.
This avoid a giant single test source file and makes it easier to create
isolated setup/teardown functions and suites.