Add status test module
The test module runner (main function in each test module) is changed
to return the failure count from the minunit test framework. This way,
if there are no failures, the return code for the main function will be
zero, i.e. success, no errors, and it will be equal to the number of
failures within the module, i.e. non-zero. This ensures the `ctest`
and/or `make test` commands will report failures. Previously, the
`ctest` and `make test` always showed PASS, even if a failure occurred
within the test module.