Remove static NULL_WRITER
This is not going to work if the developer passes a NULL for the writer.
Since parameters are passed by value, setting the writer parameter to
the NULL_WRITER's address is not going to work. The writer parameter
will revert to the previous value (NULL) when the function returns. So,
it is removed and replaced with `assert` since a NULL writer should
panic the application as something very wrong is going on. This will
work for the OUT_OF_MEMORY_WRITER because the `create` function returns
a pointer and it can be to the static struct.