Change error handling
The return codes are replaced with the `is_ok` and `is_error` functions
to determine if a function completed successfully or not. This is
cleaner and less verbose than _also_ returning a Success/Failure code
for some functions and not others. This also appears to be more
consistent with the `mpack` library implementation and other open source
projects, i.e. cairo.