Change creation of writer for LabVIEW compatibility
A handle is passed to the `create` function for a writer instead of
returning a pointer. LabVIEW cannot process pointers that are returned
from functions. Instead, a "handle", pointer to a pointer, is passed as
a parameter and populated with the address. This address is converted to
an integer in LabVIEW, which acts as a handle.