Add check for OK status before completing function
This is similar to a conventional LabVIEW VI implementation, where the
functionality is wrapped in a No Error/Error case structure. If an error
occurred from a previous VI/function call, then the body of the current
calling VI/function is _not_ executed. Instead, the error is just
passed, or propogated onto the next operation, and default values are
returned. The check is implemented just for the reader module at the
moment. TODOs have been added for the writer module.

Adding this check will ensure a "fail early" error handling mechanism.