Add "standard" error handling to writer module
A check on mpack function success is added to each function in the
public API of the writer module where appropriate. Each function has
an`if` block that skips execution if an error occurred previously. This
is to be consistent with the reader module and this is known as the
"standard" error handling functionality of the library.