Fix mpack error messages to be more human
The `mpack_error_to_string` function is kind of useless. It just
converts the mpack_error_t value to a string, for example
`mpack_error_io` would get converted to "mpack_error_io". It does not
include the reason for the error or a message. So, a helper function is
added that returns a message that provides some explanation for the
error.