Change build instructions for NI Linux RT
The `.cproject` and `.project` files are now included in the git
repository. This makes it much easier to import the project into NI
Eclipse and build for the x86_64 NI Linux RT target. The instructions
have been changed to use the new project files for Eclipse and the old
ones have been moved to a Gist.