Fix crashes with tests on Windows
The tests would crash on windows for the writer module because of
assertions that were enabled with mpack in debug mode. Thus, debug mode
is turned off for mpack under all build configurations for the shared
library and the tests. This eliminates assertions that crash but it
could lead to stealth bugs. More tests should be added to avoid bugs
slipping through with the debug mode off.